Fix more warnings

+++ src/main/java/com/khronodragon/bluestone/util/Base65536.java
@@ -91,6 +91,6 @@ public class Base65536 {
                 }
             }
         }
-        return ArrayUtils.toPrimitive(bytes.stream().map(i -> Byte.valueOf((byte) (int) i)).toArray(Byte[]::new));
+        return ArrayUtils.toPrimitive(bytes.stream().map(i -> ((byte) (int) i)).toArray(Byte[]::new));
     }
 }
